Semester 4 (Kemunduran Akademis)/Tugas/BASDAT/BASDAT - REPO TK 3 DAN TK 4 - KELOMPOK C6/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
gunicorn==22.0.0
idna==3.7
packaging==24.0
psycopg==3.1.18
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0